Creates common class for differentialDrive robots.
