+++ albertReacher/resources/albertRobot.py
@@ -5,10 +5,10 @@ import math
 from urdfpy import URDF
 import numpy as np
 
-from urdfCommon.abstractRobot import AbstractRobot
+from urdfCommon.diffDriveRobot import DiffDriveRobot
 
 
-class AlbertRobot(AbstractRobot):
+class AlbertRobot(DiffDriveRobot):
     def __init__(self):
         n = 9
         fileName = os.path.join(os.path.dirname(__file__), 'albert.urdf')
@@ -16,112 +16,14 @@ class AlbertRobot(AbstractRobot):
         self._l = 0.494
         super().__init__(n, fileName)
 
+
     def setJointIndices(self):
         self.urdf_joints = [10, 11, 14, 15, 16, 17, 18, 19, 20]
         self.robot_joints = [23, 24, 7, 8, 9, 10, 11, 12, 13]
-
-    def n(self):
-        return self._n
-
-    def reset(self, pos=None, vel=None):
-        if hasattr(self, "robot"):
-            p.resetSimulation()
-        baseOrientation = p.getQuaternionFromEuler([0, 0, pos[2]])
-        self.robot = p.loadURDF(
-            fileName=self.fileName,
-            basePosition=[pos[0], pos[1], 0.05],
-            baseOrientation=baseOrientation,
-            flags=p.URDF_USE_SELF_COLLISION_EXCLUDE_PARENT,
-        )
-        # print(self.getIndexedJointInfo())
-        # set castor wheel friction to zero
-        for i in [21, 22]:
-            p.setJointMotorControl2(
-                self.robot,
-                jointIndex=i,
-                controlMode=p.VELOCITY_CONTROL,
-                force=0.0
-            )
-        # set base velocity
-        v = np.zeros(2)
-        v[0] = vel[0] + vel[1]
-        v[1] = vel[0] - vel[0]
-        for i in range(2, self._n):
-            p.resetJointState(
-                self.robot,
-                self.robot_joints[i],
-                pos[i + 1],
-                targetVelocity=vel[i],
-            )
-        self.updateState()
-        self.apply_vel_action_wheels(v)
-        self.apply_vel_action(vel)
-        self.state[-2:] = v
-        self._vels_int = np.concatenate((self.state[-2:], self.state[13:20]))
+        self.castor_joints = [21, 22]
 
     def setAccLimits(self):
         accLimit = np.array([1.0, 1.0, 15.0, 7.5, 10.0, 12.5, 15.0, 20.0, 20.0])
         self._limitAcc_j[0, :] = -accLimit
         self._limitAcc_j[1, :] = accLimit
-
-    def apply_torque_action(self, torques):
-        for i in range(2, self._n):
-            p.setJointMotorControl2(self.robot, self.robot_joints[i],
-                                        controlMode=p.TORQUE_CONTROL,
-                                        force=torques[i])
-
-    def apply_acc_action(self, accs, dt):
-        self._vels_int += dt * accs
-        self.apply_base_velocity(self._vels_int)
-        self.apply_vel_action(self._vels_int)
-
-    def apply_vel_action_wheels(self, vels):
-        for i in range(2):
-            p.setJointMotorControl2(self.robot, self.robot_joints[i],
-                                        controlMode=p.VELOCITY_CONTROL,
-                                        targetVelocity=vels[i])
-
-    def apply_base_velocity(self, vels):
-        vel_left = (vels[0] - 0.5 * self._l * vels[1]) / self._r
-        vel_right = (vels[0] + 0.5 * self._l * vels[1]) / self._r
-        wheelVels = np.array([vel_right, vel_left])
-        self.apply_vel_action_wheels(wheelVels)
-
-    def apply_vel_action(self, vels):
-        for i in range(2, self._n):
-            p.setJointMotorControl2(self.robot, self.robot_joints[i],
-                                        controlMode=p.VELOCITY_CONTROL,
-                                        targetVelocity=vels[i])
-
-    def updateState(self):
-        # Get Base State
-        linkState = p.getLinkState(self.robot, 0, computeLinkVelocity=0)
-        posBase = np.array(
-            [
-                linkState[0][0],
-                linkState[0][1],
-                p.getEulerFromQuaternion(linkState[1])[2],
-            ]
-        )
-        velWheels = p.getJointStates(self.robot, self.robot_joints)
-        v_right = velWheels[0][1]
-        v_left = velWheels[1][1]
-        vf = np.array([0.5 * (v_right + v_left) * self._r, (v_right - v_left) * self._r / self._l])
-        Jnh = np.array([[np.cos(posBase[2]), 0], [np.sin(posBase[2]), 0], [0, 1]])
-        velBase = np.dot(Jnh, vf)
-        # Get Joint Configurations
-        joint_pos_list = []
-        joint_vel_list = []
-        for i in range(2, self._n):
-            pos, vel, _, _ = p.getJointState(self.robot, self.robot_joints[i])
-            joint_pos_list.append(pos)
-            joint_vel_list.append(vel)
-        joint_pos = np.array(joint_pos_list)
-        joint_vel = np.array(joint_vel_list)
-
-        # forward and rotational velocity
-        vf = np.array([np.sqrt(velBase[0] ** 2 + velBase[1] ** 2), velBase[2]])
-
-        # Concatenate position[0:10], velocity[0:10], vf[0:3]
-        self.state = np.concatenate((posBase, joint_pos, vf, joint_vel, velBase))
